requirements.txt
# Social Media Post Creator - Python Dependencies

# Environment
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0
httpx>=0.27.0

# AI/LLM
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Web framework (API)
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
websockets>=12.0

# Image processing
Pillow>=10.0.0

# Utilities
pydantic>=2.0.0

# Video/Audio Processing
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0

# Job Queue
celery>=5.3.0
redis>=5.0.0
